requirements.txt
streamlit
openai
python-docx
pandas
numpy
plotly